Add: Methods for displaying related entities

+++ Hache.Server/Entities/Venta.cs
@@ -5,8 +5,8 @@
         public int ID_Venta { get; set; }
         public Usuario Usuario { get; set; } = new Usuario();
         public DateTime Fecha { get; set; }
-        public float Subtotal { get; set; }
-        public float Total { get; set; }
+        public decimal Subtotal { get; set; }
+        public decimal Total { get; set; }
         public bool EsPedidosYa { get; set; }
         public Local Local { get; set; } = new Local(); 
         public Venta() { }
